Adjusted demo message loop handler. Added additional documentation to AutoMarshal. Clean up.

+++ core/src/main/java/com/etdon/winj/input/InputSender.java
@@ -71,7 +71,7 @@ public class InputSender {
                     .build()
                     .createMemorySegment(this.arena), 0, inputArray, Input.INPUT.byteSize(), Input.INPUT.byteSize());
 
-            return (int) this.nativeCaller.call(
+            return this.nativeCaller.call(
                     SendInput.builder()
                             .inputArray(inputArray)
                             .inputCount(2)
